+++ src/components/Card/CardContent.tsx
@@ -2,6 +2,12 @@ import React from 'react';
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
 
 export interface CardContentProps extends React.HTMLAttributes<HTMLDivElement> {
+    /**
+     * Usually `CardContent` uses all the remaining space inside a card.
+     * This behaviour can be changed by this option, so that the content section is not vertically grown or shrinked.
+     * You may use this then you have multiple content sections in one card but not all those sections should be grown or shrinked regarding the remaining space.
+     */
+    noFlexHeight?: boolean;
 }
 
 /**
@@ -11,12 +17,17 @@ export interface CardContentProps extends React.HTMLAttributes<HTMLDivElement> {
 function CardContent({
     children,
     className='',
+    noFlexHeight,
     ...otherProps
 }: CardContentProps) {
     return (
         <div
             {...otherProps}
-            className={`${eccgui}-card__content ` + className}
+            className={
+                `${eccgui}-card__content` +
+                (noFlexHeight ? ` ${eccgui}-card__content--noflexheight` : "") +
+                (!!className ? ` ${className}` : "")
+            }
         >
             {children}
         </div>